chore: cleanup code structure
